+++ app/src/main/kotlin/com/simplemobiletools/calendar/pro/activities/MainActivity.kt
@@ -267,11 +267,11 @@ class MainActivity : SimpleActivity(), RefreshRecyclerViewListener {
             return true
         }
 
-        val eventIdToOpen = intent.getIntExtra(EVENT_ID, 0)
+        val eventIdToOpen = intent.getLongExtra(EVENT_ID, 0)
         val eventOccurrenceToOpen = intent.getIntExtra(EVENT_OCCURRENCE_TS, 0)
         intent.removeExtra(EVENT_ID)
         intent.removeExtra(EVENT_OCCURRENCE_TS)
-        if (eventIdToOpen != 0 && eventOccurrenceToOpen != 0) {
+        if (eventIdToOpen != 0L && eventOccurrenceToOpen != 0) {
             Intent(this, EventActivity::class.java).apply {
                 putExtra(EVENT_ID, eventIdToOpen)
                 putExtra(EVENT_OCCURRENCE_TS, eventOccurrenceToOpen)
@@ -290,7 +290,7 @@ class MainActivity : SimpleActivity(), RefreshRecyclerViewListener {
                     // intents like content://com.android.calendar/events/1756
                     val eventId = uri.lastPathSegment
                     val id = dbHelper.getEventIdWithLastImportId(eventId)
-                    if (id != 0) {
+                    if (id != 0L) {
                         Intent(this, EventActivity::class.java).apply {
                             putExtra(EVENT_ID, id)
                             startActivity(this)
